fixes #137 - more info for Java caller about Python callback failure
Instead of just saying 'x' if a Python callback fails, we will throw an
informative Java exception that includes the Python backtrace.

We also add a new configuration flag, propagate_java_exceptions. If this
is turned on then if Python callback raises a Py4JJavaError, the nested
Java exception will be rethrown on the Java side instead of a regular
Py4JException. This lets Python implement interfaces where it is
important that it throws certain Java exception types.