Add support for auth tokens to gateway and callback connections.
This change adds authentication to both the gateway and callback
servers, in the form of a shared authentication token.

The generation and distribution of tokens is left for applications
to do. The library just provides the means for enforcing the tokens.
The only exception is when using the gateway server command line,
in which case it will generate a token when requested.

By default, like previously, there is no authentication.

The auth protocol is very, very simple: when auth is enabled, the
servers expect the first line of content from the client connection
to be the auth token. If the token doesn't match, the server will
shut down the connection. Note this means that the token is sent
in plain text over the connection, meaning that for security it
requires the use of a local socket, or SSL for remote sockets.

Fixes #268