Partially fixes #320 -- copies @bartdag's code to return the process.
The Popen object is always created on the JavaGateway with name java_process,
with a value of None when no Popen is available because the JavaGateway is
attaching to a preexisting Java process.
It can thus be referenced later, including to terminate it by sending
the end-of-line character it is waiting for.

Adds tests using the Popen member to test terminating the Java subprocess.

The test waits an arbitrarily-chosen fraction of a second
before asserting that the Java process has terminated.