fixes #507 -- Fix GitHub workflow badge
See the linked GitHub issue from the previous version of the badge:
https://github.com/badges/shields/issues/8671

In summary, this badge was changed in a breaking fashion such that the
badge was just linking to the github issue rather than showing the data.
Updating the url resolves this.

I also updated to using py4j as the GitHub user.